requirements.txt
numpy
pandas
sklearn
tensorflow
keras
pydataset
pyyaml
statsmodels
tqdm
keras-rectified-adam
keras-lookahead
seaborn
pmdarima